Some bug fixes

+++ src/routes/tweetRoutes.ts
@@ -1,5 +1,6 @@
 import { Router } from "express";
 import { PrismaClient } from "@prisma/client";
+import req from "../middlewares/tokenVerification"
 
 const router = Router();
 const prisma = new PrismaClient();
@@ -23,7 +24,32 @@ router.post("/", async (req, res) => {
     }
 });
 
-//. List Tweets
+//. List Tweets of a user
+router.get("/user/all", async (req, res) => {
+    try {
+        // We can also include user object(or only some information) from whom the tweet is related 
+        // (with userID and tweet relation defined in schema)
+        const alltweets = await prisma.tweet.findMany({
+            where: {
+                userID: req.user
+            },
+            include: {
+                user: {
+                    select: {
+                        username: true,
+                        image: true
+                    }
+                }
+            }
+        });
+        res.json(alltweets);
+    } catch (error) {
+        console.log(error);
+        res.json({ "error": error });
+    }
+});
+
+//. Get all tweets
 router.get("/all", async (req, res) => {
     try {
         // We can also include user object(or only some information) from whom the tweet is related 
@@ -32,10 +58,8 @@ router.get("/all", async (req, res) => {
             include: {
                 user: {
                     select: {
-                        id: true,
-                        name: true,
                         username: true,
-                        image: true,
+                        image: true
                     }
                 }
             }
@@ -51,7 +75,7 @@ router.get("/all", async (req, res) => {
 router.get("/:id", async (req, res) => {
     const { id } = req.params;
     try {
-        const tweet = await prisma.tweet.findUnique({ where: { id: id }, include: { user: true } });
+        const tweet = await prisma.tweet.findUnique({ where: { id: id } });
         if (!tweet) res.send("No such tweet exist");
         else res.json(tweet);
     } catch (error) {
@@ -63,13 +87,22 @@ router.get("/:id", async (req, res) => {
 //. Update Tweet
 router.put("/:id", async (req, res) => {
     const { id } = req.params;
-    const { content} = req.body;
+    const { content } = req.body;
     try {
-        const updateTweet = await prisma.tweet.update({
-            where: { id },
-            data: { content }
+        const checkID = await prisma.tweet.findUnique({
+            where: {
+                id,
+                userID: req.user
+            }
         })
-        res.json(updateTweet);
+        if (checkID) {
+            const updateTweet = await prisma.tweet.update({
+                where: { id },
+                data: { content }
+            })
+            res.json(updateTweet);
+        }
+        res.json({ "error": "You are not allowed to update" })
     } catch (error) {
         console.log(error);
         res.json({ "error": "Something bad happened !" });
@@ -80,8 +113,17 @@ router.put("/:id", async (req, res) => {
 router.delete("/:id", async (req, res) => {
     const { id } = req.params;
     try {
-        await prisma.tweet.delete({ where: { id: id } });
-        res.send("Successful deletion");
+        const checkID = await prisma.tweet.findUnique({
+            where: {
+                id,
+                userID: req.user
+            }
+        })
+        if (checkID) {
+            await prisma.tweet.delete({ where: { id: id } });
+            res.send("Successful deletion");
+        }
+        res.json({ "json": "You are not allowed !" });
     } catch (error) {
         console.log(error);
         res.json({ "error": "Something bad happened !" });
